requirements.txt
blessed==1.22.0
editor==1.6.6
inquirer==3.4.1
python-dotenv==1.1.1
readchar==4.2.1
runs==1.2.2
wcwidth==0.2.14
xmod==1.8.1
